Sine function
- Sine function
- Change return type

+++ Interpreter/src/Expression.java
@@ -1,5 +1,5 @@
 
 public interface Expression {
 //	public int interpret(Map<String,Expression> variables);
-	public int interpret();
+	public double interpret();
 }
