+++ Interpreter/src/Plus.java
@@ -7,7 +7,7 @@ public class Plus implements Expression{
     }
  
     @Override
-    public int interpret()  { 
+    public double interpret()  { 
         return leftOperand.interpret() + rightOperand.interpret();
     }
